+++ scripts/website.js
@@ -133,7 +133,6 @@ function getVersions() {
     filteredTags.push([0,0,0]);
   }
 }
-<<<<<<< HEAD
 
 /**
  * Stringify a semver number array
@@ -156,30 +155,6 @@ function getLatestVersion() {
   return { listed: stringifySemverNumber(filteredTags[0]), path: '' };
 }
 
-=======
-
-/**
- * Stringify a semver number array
- * @param {number[]} arr The array to stringify
- * @param {boolean} dotX If "true", return "5.X" instead of "5.5.5"
- * @returns 
- */
-function stringifySemverNumber(arr, dotX) {
-  if (dotX) {
-    return `${arr[0]}.x`;  
-  }
-  return `${arr[0]}.${arr[1]}.${arr[2]}`;
-}
-
-/** 
- * Get the latest version available
- * @returns {Version}
- */
-function getLatestVersion() {
-  return { listed: stringifySemverNumber(filteredTags[0]), path: '' };
-}
-
->>>>>>> 6.x
 /**
  * Get the latest version for the provided major version
  * @param {number} version major version to search for
@@ -257,33 +232,21 @@ try {
 
 const docsFilemap = require('../docs/source/index');
 const files = Object.keys(docsFilemap.fileMap);
-<<<<<<< HEAD
 // api explicitly imported for specific file loading
 const apiReq = require('../docs/source/api');
-=======
->>>>>>> 6.x
 
 const wrapMarkdown = (md, baseLayout, versionedPath) => `
 extends ${baseLayout}
 
 append style
-<<<<<<< HEAD
   link(rel="stylesheet", href="${versionedPath}/docs/css/inlinecpc.css")
   script(type="text/javascript" src="${versionedPath}/docs/js/native.js")
-=======
-  link(rel="stylesheet", href="#{versions.versionedPath}/docs/css/inlinecpc.css")
-  script(type="text/javascript" src="#{versions.versionedPath}/docs/js/native.js")
->>>>>>> 6.x
   style.
     p { line-height: 1.5em }
 
 block content
   <a class="edit-docs-link" href="#{editLink}" target="_blank">
-<<<<<<< HEAD
     <img src="${versionedPath}/docs/images/pencil.svg" />
-=======
-    <img src="#{versions.versionedPath}/docs/images/pencil.svg" />
->>>>>>> 6.x
   </a>
   :markdown
 ${md.split('\n').map(line => '    ' + line).join('\n')}