+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+7.1.1 / 2023-05-10
+==================
+ * fix(document): handle set() from top-level underneath a map of mixed #13386
+ * fix: don't modify passed options object to `createConnection()` #13376
+ * types: make lean() not clobber result type for updateOne(), etc. #13389 #13382
+ * types: handle union types in FlattenMaps #13368 #13346 [Jokero](https://github.com/Jokero)
+ * types(document): correct return type for Model.prototype.deleteOne(): promise, not query #13367 #13223
+ * types: update document.d.ts $set function params to match set #13304 [jeffersonlipsky](https://github.com/jeffersonlipsky)
+ * docs: add excludeIndexes to the guide schema options list #13377 #13287
+ * docs: fix broken "fork me" on home page #13336
+
+6.11.1 / 2023-05-08
+===================
+ * fix(query): apply schema-level paths before calculating projection for findOneAndUpdate() #13348 #13340
+ * fix: add SUPPRESS_JEST_WARNINGS environment variable to hide jest warnings #13384 #13373
+ * types(model): allow overwriting expected param type for bulkWrite() #13292 [hasezoey](https://github.com/hasezoey)
+
 6.11.0 / 2023-05-01
 ===================
  * feat: upgrade to mongodb 4.16.0 for Deno+Atlas connection fix #13337 #13075